refactor: move setter generation logic to a separate module for better organization
